# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for managing Pointlings, users, items, and XP system",
        "title": "Pointlings API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/items": {
            "get": {
                "description": "Get a list of items with optional filtering by category, rarity, and slot",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "List available items",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by item category (e.g. COSMETIC, CONSUMABLE)",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by item rarity (e.g. COMMON, RARE, EPIC)",
                        "name": "rarity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by item slot (e.g. HEAD, BODY, ACCESSORY)",
                        "name": "slot",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Item"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid category/rarity/slot",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new item in the catalog (admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Create a new item",
                "parameters": [
                    {
                        "description": "Item creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Item"
                        }
                    },
                    "400": {
                        "description": "Invalid request body/parameters",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/items/{item_id}": {
            "get": {
                "description": "Get detailed information about a specific item",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get item details",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Item"
                        }
                    },
                    "400": {
                        "description": "Invalid item ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Item not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/pointlings/{pointling_id}/items": {
            "get": {
                "description": "List all items owned by a pointling with optional equipped filter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Get pointling's inventory",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by equipped status",
                        "name": "equipped",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Item"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid pointling ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pointlings/{pointling_id}/items/{item_id}": {
            "post": {
                "description": "Add an item to a pointling's inventory (requires meeting level requirements)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Acquire an item for a pointling",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Item"
                        }
                    },
                    "400": {
                        "description": "Invalid pointling/item ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Level requirement not met",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pointling/item not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Item already owned",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pointlings/{pointling_id}/items/{item_id}/equip": {
            "patch": {
                "description": "Equip or unequip an item for a pointling",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Items"
                ],
                "summary": "Toggle item equipped status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Pointling ID",
                        "name": "pointling_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Item ID",
                        "name": "item_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Toggle equipped request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.toggleEquippedRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid pointling/item ID or request body",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Pointling does not own this item",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "handlers.createItemRequest": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string",
                    "example": "hat_001"
                },
                "category": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ItemCategory"
                        }
                    ],
                    "example": "COSMETIC"
                },
                "name": {
                    "type": "string",
                    "example": "Cool Hat"
                },
                "price_points": {
                    "type": "integer",
                    "example": 100
                },
                "rarity": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ItemRarity"
                        }
                    ],
                    "example": "RARE"
                },
                "slot": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ItemSlot"
                        }
                    ],
                    "example": "HEAD"
                },
                "unlock_level": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
...
        "handlers.toggleEquippedRequest": {
            "type": "object",
            "properties": {
                "equipped": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.Item": {
            "type": "object",
            "properties": {
                "asset_id": {
                    "type": "string"
                },
                "category": {
                    "$ref": "#/definitions/models.ItemCategory"
                },
                "item_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "price_points": {
                    "type": "integer"
                },
                "rarity": {
                    "$ref": "#/definitions/models.ItemRarity"
                },
                "slot": {
                    "$ref": "#/definitions/models.ItemSlot"
                },
                "unlock_level": {
                    "type": "integer"
                }
            }
        },
        "models.ItemCategory": {
            "type": "string",
            "enum": [
                "ACCESSORY",
                "FEATURE"
            ],
            "x-enum-varnames": [
                "CategoryAccessory",
                "CategoryFeature"
            ]
        },
        "models.ItemRarity": {
            "type": "string",
            "enum": [
                "COMMON",
                "RARE",
                "EPIC",
                "LEGENDARY"
            ],
            "x-enum-varnames": [
                "RarityCommon",
                "RarityRare",
                "RarityEpic",
                "RarityLegendary"
            ]
        },
        "models.ItemSlot": {
            "type": "string",
            "enum": [
                "HAT",
                "SHOES",
                "FACE",
                "WINGS"
            ],
            "x-enum-varnames": [
                "SlotHat",
                "SlotShoes",
                "SlotFace",
                "SlotWings"
            ]
        },
...
    },
...
}